requirements.txt
flask==3.1.1
flask_sqlalchemy==3.1.1
flask_login==0.6.3
jinja2==3.1.6
Flask-Migrate==4.1.0
Flask-WTF==1.2.2
email_validator==2.2.0
python-dotenv==1.1.1
humanize==4.13.0
pillow==11.3.0
pathvalidate==3.3.1